refactor to make GUI run ytmusic_deleter modules directly instead of bundling CLI as executable
