+++ ytmusic_deleter/_version.py
@@ -1,1 +1,1 @@
-__version__ = "3.6.3.dev25+gb8810df.d20260320"
+__version__ = "3.6.3.dev28+g91793d0.d20260321"
